Tested the logging functionality as a response to some errors detected
I found no error using the code on my computer (windows) however it can be due to some directory name conventions that are different between windows and unix (which was the one that didn't work for Leo). There are some new logged files in the ./stories and ./stories2 folders and none are essential. If you want you can delete them.